polished config & updated readme

+++ src/interactions/buttons/remove.ts
@@ -5,6 +5,12 @@ const remove: Interaction = {
     customId: `remove`,
     run: async (client, interaction) => {
         if(!(interaction instanceof ButtonInteraction)) return;
+        
+        if(interaction.guild?.ownerId != interaction.user.id) {
+            interaction.reply({ content: client.i18n.__("utils.noPermissions"), ephemeral: true });
+            return;
+        }
+
         interaction.message.delete();
     }
 }
